+++ changelog.md
@@ -1,10 +1,13 @@
 # CHANGELOG
 
-## [0.4.1] 2020-07-23
+## [0.4.1] 2020-07-25
 ### ADDED
+- Ability to push a token to behind other tokens in stack with keystroke.
+- Reworked code for more consistant tooltip and mouse-click behavior.
 ### BUGFIXES
 - Keyboard shortcut would continue to trigger cycle when user opened character sheet.
-- Third click would not invoke a cycle on a stack of tokens, usually after F5 pressed.
+- Tooltip would disappear when clicking tokens in stack.
+- Clicking on the token would not always cycle the stack.
 
 ## [0.4.0] 2020-07-20
 ### ADDED